+++ src/Record/Traits.php
@@ -76,6 +76,14 @@ use GeoIp2\Util;
  * @property-read string|null $organization The name of the organization associated
  * with the IP address. This attribute is only available from the City and
  * Insights web services and the GeoIP2 Enterprise database.
+ * @property-read string|null $mobileCountryCode The [mobile country code
+ * (MCC)](https://en.wikipedia.org/wiki/Mobile_country_code) associated with
+ * the IP address and ISP. This property is available from the City and
+ * Insights web services and the GeoIP2 Enterprise database.
+ * @property-read string|null $mobileNetworkCode The [mobile network code
+ * (MNC)](https://en.wikipedia.org/wiki/Mobile_country_code) associated with
+ * the IP address and ISP. This property is available from the City and
+ * Insights web services and the GeoIP2 Enterprise database.
  * @property-read float|null $staticIpScore An indicator of how static or
  * dynamic an IP address is. This property is only available from GeoIP2
  * Precision Insights.
@@ -130,6 +138,8 @@ class Traits extends AbstractRecord
         'isResidentialProxy',
         'isSatelliteProvider',
         'isTorExitNode',
+        'mobileCountryCode',
+        'mobileNetworkCode',
         'network',
         'organization',
         'staticIpScore',